line number should be 32 bits
git-svn-id: https://uncrustify.svn.sourceforge.net/svnroot/uncrustify/trunk@1175 74908cc9-670d-0410-99d4-ca74d6b17880

+++ src/uncrustify_types.h
@@ -294,7 +294,7 @@ struct cp_data
 
    int                lang_flags; // LANG_xxx
 
-   UINT16             line_number;
+   UINT32             line_number;
    UINT16             column;  /* column for parsing */
 
    /* stuff to auto-detect line endings */
